Enable :lax_htmlblock when parsing MD
Although the MD standard specifies that inline HTML blocks must be
surrounded by blank lines before and after the block, the Markdown.pl
parser does not honor this. We use the `lax_htmlblock` flag to mimic the
output of the official parser, so now blocks like

	This is a paragraph
	<pre>
		this is an inline html block
	</pre>
	This is another paragraph

are considered inline blocks.