Test for commit range with invalid SHA.
This and the test to not link a range using .. seem a little off to me.
Should we still link these to individual commits? Maybe so if at least
part of it is a valid SHA. It seems more likely that someone is
accidentally hitting on range-ish syntax (by
typing..something..like...this) than valid SHAs.