Leave link URLs using the input SHAs
No need to be clever with shortening the SHAs used to make these links.
I was actually just about to pull this shortening behavior into the
regular commit-linking method (`replace_bare_commit_mentions`), but
decided instead to go the other way and take what was used as input,
only shortening the display text.