Bringing back the explicit range-dot regex
This reverts 32f7672 and 755df65.

Tighter regexes are awesome, but not at the expense of correct
functionality. @bkeepers pointed out we probably don't want to SHA-link
something like `somesubdomain.abcd0123.com`, and I agree with him.

Now there's a test that supports this, but there's just one. If there
are other SHA-like things we definitely want to not turn into SHA links,
we should make sure there are tests to cover those cases.