Clarify state handling in GitHub::HTML::Pipeline
Previously we were jumping through hoops to manage initial context
created with a pipeline, and then context that is passed in for
every `call` of a pipeline.  It was nasty, error prone, and I didn't
see a case where we were clearing out old state from previous runs
of pipeline.

Here I'm introducing a `default_context` which is passed to the
creation of a Pipeline, and the normal `context` that is passed into
call.  `default_context` gets merged onto `context`, so it always
wins, and both are frozen so they cannot be changed once set.

It still feels kinda crummy that we reuse these Pipeline instances
for our processing, as we don't have a good assurance that state
  is cleaned up between runs.  We basically have to very carefully
  code for it, hence the gradual move over time to freeze things like
  Result and context.