Remove RUBY_VERSION conditionals from gemspec
Remove the useless conditionals added in ae1f3c828cccd2b5ebf0678ea3edc2bcca10c9f3
since gemspec is turned into a static one at gem publish time, so the
version constraints will be adjusted to whatever Ruby version the gem
was *published* with, not the one that will eventually run the gem.

If someone is still running Ruby 1.8 or 1.9.2, it's their responsibility
to adjust their project's Gemfile to prevent Nokogiri or Active Support
upgrading to an incompatible version.