Whitelist various inline semantic/formatting tags
Mostly I want <abbr>, but the rest of these should also be whitelisted.

What goes in LIMITED.elements seems to be a bit subjective, but I think
that <mark> and <abbr> at least are reasonably justified. <dfn> is
fairly obscure, and <small>, <bdo>, <cite>, <span>, <time> and <wbr> a
bit different from the rest and probably not justified, so I left them
out.

In the absence of styling, <span> is quite useless, but <div>’s there
already, and people have asked for <span> (#183), so I figured, why not?

Fixes #183.