Add 'force_encoding' option to MarkdownFilter
This is to fix an open issue with a dependant project:
https://github.com/alphagov/govuk-developer-docs/issues/1821

HTML::Pipeline uses CommonMarker as its markdown parser in the
MarkdownFilter. CommonMarker throws an exception if it encounters
any non-UTF-8 content:
https://github.com/gjtorikian/commonmarker/pull/10/files#diff-5850b052ec7a5e6bd8b51bc53465146aR8

We would like to be able to pass a `force_encoding: "UTF-8"`
parameter to the context, which can be used to force encoding
prior to calling CommonMarker, so that the exception is avoided.